Navigation, DeviceToSave

+++ MockupForProject/app/src/main/java/com/hkr/mockupforproject/data/SavedDevice.kt
@@ -7,7 +7,7 @@ data class SavedDevice(
     var model: String = "Not defined",
     var recommendation: String = "Not defined",
     var nearbyTowers: List<CellTower> = emptyList(),
-    var imei: Int = 0,
+    var imei: Long = 0,
     var priority: Int = 0
 ) {
